Handle CSS @imports
When a CSS file changes, looks up in goog.cssom/getAllCssStyleSheets to
verify it's in the HTML file, then repeatedly checks .-parentStylesheet
to locate the top-level link.

Also switches from removing the old link after a 300ms timeout to
removing it after the new link's load event fires. This is 10-20x faster.