Adds a proposed emphatic note to the Vim help
Yes, starting the repl first _is_ implied in these docs and the repl-env
docstring.

The trouble for me came in that it seems like Piggieback sometimes will
and sometimes won't start up a cljs-repl on its own given a repl-env. I
got to assuming that `:Piggieback (figwheel.main.api/repl-env "dev")` in
vim was roughly equivalent to starting the cljs-repl in the clj-repl.

I'm presuming others might run into this confusion and want to make the
right thing a little more explicit.