Fix typo and javascript code sample (#232)
I believe the top line of the last javascript code block will not work with `{goog = window}` and instead needs to be `{goog: window}`. At least that's how I had to do it when testing in the browser console. I'm guessing it's just a typo, bu I could be misunderstanding something.