Change note about CLI Tools on Windows (#229)
The CLI Tools are available on Windows now in an alpha state. I've used the CLI Tools on Windows for the Figwheel tutorial and it's been working fine. Since it's currently in an alpha state I don't know if you want to remove the section "On Windows" or just change it to something like I did.